+++ comedores/views/dupla.py
@@ -1,0 +1,37 @@
+from django.contrib import messages
+from django.contrib.auth.mixins import LoginRequiredMixin
+from django.shortcuts import redirect
+from django.views.generic import ListView
+
+from comedores.models import Comedor
+from comedores.services.comedor_service import ComedorService
+from duplas.dupla_service import DuplaService
+
+
+# TODO: Sacar de la vista de comedores
+class AsignarDuplaListView(LoginRequiredMixin, ListView):
+    model = Comedor
+    template_name = "comedor/asignar_dupla_form.html"
+
+    def get_context_data(self, **kwargs):
+        data = super().get_context_data(**kwargs)
+        comedor = ComedorService.get_comedor(self.kwargs["pk"])
+        duplas = DuplaService.get_duplas_by_estado_activo()
+        data["comedor"] = comedor
+        data["duplas"] = duplas
+        return data
+
+    def post(self, request, *args, **kwargs):
+        dupla_id = request.POST.get("dupla_id")
+        comedor_id = self.kwargs["pk"]
+
+        if dupla_id:
+            try:
+                ComedorService.asignar_dupla_a_comedor(dupla_id, comedor_id)
+                messages.success(request, "Dupla asignada correctamente.")
+            except Exception as e:
+                messages.error(request, f"Error al asignar la dupla: {e}")
+        else:
+            messages.error(request, "No se seleccionó ninguna dupla.")
+
+        return redirect("comedor_detalle", pk=comedor_id)
